Add much more robust version, still not perfect
