+++ .github/workflows/reusable_smoke_tests.yml
@@ -160,7 +160,7 @@ jobs:
             -e MEMGRAPH_LAST_DOCKERHUB_IMAGE="$MEMGRAPH_LAST_DOCKERHUB_IMAGE" \
             -e MEMGRAPH_NEXT_DOCKERHUB_IMAGE="$MEMGRAPH_NEXT_DOCKERHUB_IMAGE" \
             "$CONTAINER_NAME" \
-            bash -c "cd /mage/smoke-release-testing && ./test.bash"
+            bash -c "cd /mage/smoke-release-testing && ./test_single_mage.bash"
 
       - name: Clean up inner Docker state
         if: always()
