+++ smoke-release-testing/test_single_mage.bash
@@ -3,11 +3,9 @@ SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source "$SCRIPT_DIR/utils.bash"
 
 # TODO(gitbuda): Measure the total execution time, it should be under ~10s.
-# TODO(gitbuda): Test docker compose.
-# TODO(gitbuda): memgraph vs mage -> docker tests assume mage while helm is memgraph -> RESOLVE.
 
 # NOTE: 1st arg is how to pull LAST image, 2nd arg is how to pull NEXT image.
-spinup_and_cleanup_memgraph_dockers DockerHub RC
+spinup_and_cleanup_memgraph_dockers DockerHub DockerHub
 echo "Waiting for memgraph to initialize..."
 wait_for_memgraph $MEMGRAPH_DEFAULT_HOST $MEMGRAPH_LAST_DATA_BOLT_PORT
 wait_for_memgraph $MEMGRAPH_DEFAULT_HOST $MEMGRAPH_NEXT_DATA_BOLT_PORT
@@ -50,9 +48,3 @@ test_or_expression_for_labels $MEMGRAPH_DEFAULT_HOST $MEMGRAPH_NEXT_DATA_BOLT_PO
 # NOTE: If the testing container is NOT restarted, all the auth test have to
 # come after all tests that assume there are no users.
 test_impersonate_user $MEMGRAPH_DEFAULT_HOST $MEMGRAPH_NEXT_DATA_BOLT_PORT
-
-# k8s is a special case, because it requires extra setup.
-source $SCRIPT_DIR/k8s/run.bash
-test_k8s_single
-test_k8s_ha LAST -c # Skip cleanup because we want to recover from existing PVCs.
-test_k8s_ha NEXT -s # Skip cluster setup because that should be recovered from PVCs.
