+++ smoke-release-testing/utils.bash
@@ -72,31 +72,33 @@ wait_port() {
 wait_for_memgraph() {
   __host=$1
   __port=$2
-  __max_retries=${3:-50}
+  __max_retries=${3:-100}
   __retries=0
   while ! echo "RETURN 1;" | $MEMGRAPH_CONSOLE_BINARY --host $__host --port $__port > /dev/null 2>&1; do
-    sleep 0.1
+    sleep 0.2
     __retries=$((__retries+1))
     if [ "$__retries" -ge "$__max_retries" ]; then
       echo "wait_for_memgraph: Reached max retries ($__max_retries) for $__host:$__port"
       return 1
     fi
   done
+  return 0
 }
 
 wait_for_memgraph_coordinator() {
   __host=$1
   __port=$2
-  __max_retries=${3:-50}
+  __max_retries=${3:-100}
   __retries=0
   while ! echo "SHOW INSTANCE;" | $MEMGRAPH_CONSOLE_BINARY --host $__host --port $__port > /dev/null 2>&1; do
-    sleep 0.1
+    sleep 0.2
     __retries=$((__retries+1))
     if [ "$__retries" -ge "$__max_retries" ]; then
       echo "wait_for_memgraph_coordinator: Reached max retries ($__max_retries) for $__host:$__port"
       return 1
     fi
   done
+  return 0
 }
 
 run_memgraph_binary() {
@@ -247,35 +249,49 @@ spinup_and_cleanup_memgraph_dockers() {
 }
 
 with_kubectl_portforward() (
-    # --- 1. peel off first two required arguments ---------------------------
-    local target=$1 # svc/foo, pod/bar, deployment/baz, …
-    local map=$2    # 8080:80 or 8443 or 0.0.0.0:8080:80
-    shift 3         # “--” + user commands; NOTE: Use an array of commands
-                    #     inside a string:
-                    #     -- 'cmd1' 'cmd2' ...
-                    #     because if you use ; or && to separate
-                    #     commands, bash will treat that as one
-                    #     command + cleanup + the rest.
+    local target=$1  # svc/foo, pod/bar, deployment/baz, …
+    local map=$2     # 8080:80 or 8443 or 0.0.0.0:8080:80
+    local probe="$3" # probe to test the target process, e.g. wait_for_xyz
+    shift 4          # “--” + user commands; NOTE: Use an array of commands
+                     #     inside a string: -- 'cmd1' 'cmd2' ...
+                     #     because if you use ; or && to separate commands,
+                     #     bash will treat that as one command + cleanup + the
+                     #     rest.
 
-    # --- 2. launch port-forward in background -------------------------------
-    # Send its output to a temp file so we can inspect errors if it dies early
     local log
-    log=$(mktemp)
-    # TODO(gitbuda): port-forward sometimes failes (something is flaky) -> FIX
-    kubectl port-forward "$target" "$map" >/dev/null 2>>"$log" &
-    pf_status=$?
-    echo "$pf_status"
-    local pf_pid=$!
+    local pf_pid
+    local retries=0
+    local max_retries=5
+    while true; do
+        log=$(mktemp)
+        kubectl port-forward "$target" "$map" >/dev/null 2>>"$log" &
+        pf_pid=$!
+        # NOTE: port-forward doesn't have built-in timeout. The target process
+        # might take arbitrary time to initialize. -> The only way to know if
+        # everything is right in the shortest amount of time is to inject the
+        # probe as one of the required params.
+        # TODO(gitbuda): If this is lower (depending how much), wait_for_coordinator fails or MAIN is not elected in the cluster yet... -> FIX
+        sleep 1
+        if ! eval "$probe"; then
+          kill -9 "$pf_pid" 2>/dev/null || true
+          wait "$pf_pid" 2>/dev/null || true
+          retries=$((retries+1))
+          if [ $retries -ge $max_retries ]; then
+            echo "kubectl port-forward failed after $max_retries attempts — see $log and inspect the target process" >&2
+            exit 1
+          fi
+        else
+          break
+        fi
+    done
 
     cleanup() {
         echo "calling port forward cleanup"
-        kill "$pf_pid" 2>/dev/null || true
-        wait "$pf_pid" 2>/dev/null || true   # reap the child
-        rm -f "$log"
+        kill -9 "$pf_pid" 2>/dev/null || true
+        wait "$pf_pid" 2>/dev/null || true
     }
     trap cleanup EXIT INT TERM
 
-    # --- 3. wait until the local port is actually open ---------------------
     local lport=${map%%:*}              # leftmost number before the first ':'
     for _ in {1..50}; do                # ~10 s max (50×0.2 s)
         if nc -z 127.0.0.1 "$lport" 2>/dev/null; then break; fi
@@ -286,9 +302,7 @@ with_kubectl_portforward() (
         sleep 0.2
     done
 
-    # --- 4. hand control to the caller’s command list ----------------------
     for cmd in "$@"; do
       eval "$cmd"
     done
-    # cleanup() runs automatically thanks to the trap
 )
