+++ smoke-release-testing/validator.py
@@ -15,6 +15,17 @@ def validate_first_as_int(data, field, expected_value):
     print(f"Validation of the first {field} is OK.")
 
 
+def get_main_parser(data):
+    main_addr=None
+    for instance in data:
+        if instance["role"] == '"main"':
+            main_addr=instance["management_server"][1:-1].split(".")[0]
+    print(main_addr)
+
+def validate_number_of_results(data, expected):
+    assert len(data) == int(expected)
+
+
 def get_arguments():
     parser = argparse.ArgumentParser(
         description="Smoke Tests Validator",
@@ -33,14 +44,27 @@ def get_arguments():
         "-e", "--expected", help="Expected value", required=True
     )
 
+    subparsers.add_parser(
+        "get_main_parser", help="Get main parser"
+    )
+
+    validate_number_of_results_parser = subparsers.add_parser(
+        "validate_number_of_results", help="Get the number of results"
+    )
+    validate_number_of_results_parser.add_argument(
+        "-e", "--expected", help="Expected value", required=True
+    )
+
     return parser.parse_args()
 
 
 if __name__ == "__main__":
-
     args = get_arguments()
-
     data = read_all_csv_from_stdin()
 
     if args.action == "first_as_int":
         validate_first_as_int(data, args.field, args.expected)
+    if args.action == "get_main_parser":
+        get_main_parser(data)
+    if args.action == "validate_number_of_results":
+        validate_number_of_results(data, args.expected)
